Bugfix 2.0.x

+++ Marlin/src/feature/bedlevel/ubl/ubl.cpp
@@ -50,7 +50,7 @@
     GRID_LOOP(x, y)
       if (!isnan(z_values[x][y])) {
         SERIAL_ECHO_START();
-        SERIAL_ECHOPAIR("  M421 I", int(x), " J", int(y));
+        SERIAL_ECHOPAIR("  M421 I", x, " J", y);
         SERIAL_ECHOLNPAIR_F_P(SP_Z_STR, z_values[x][y], 4);
         serial_delay(75); // Prevent Printrun from exploding
       }
@@ -150,7 +150,7 @@
     SERIAL_ECHO_SP(7);
     LOOP_L_N(i, GRID_MAX_POINTS_X) {
       if (i < 10) SERIAL_CHAR(' ');
-      SERIAL_ECHO((int)i);
+      SERIAL_ECHO(i);
       SERIAL_ECHO_SP(sp);
     }
     serial_delay(10);
@@ -214,7 +214,7 @@
         else if (isnan(f))
           serialprintPGM(human ? PSTR("  .   ") : PSTR("NAN"));
         else if (human || csv) {
-          if (human && f >= 0.0) SERIAL_CHAR(f > 0 ? '+' : ' ');  // Space for positive ('-' for negative)
+          if (human && f >= 0.0) SERIAL_CHAR(f > 0 ? '+' : ' ');  // Display sign also for positive numbers (' ' for 0)
           SERIAL_ECHO_F(f, 3);                                    // Positive: 5 digits, Negative: 6 digits
         }
         if (csv && i < GRID_MAX_POINTS_X - 1) SERIAL_CHAR('\t');
